+++ src/libraries/gtk+2/doinst.sh
@@ -16,10 +16,12 @@ config() {
     OLD="`dirname $NEW`/`basename $NEW .new`"
     # If there's no config file by that name, mv it over:
     if [ ! -r $OLD ]; then
-      mv $NEW $OLD
+      if [ -f $NEW ]; then
+        mv $NEW $OLD
+      fi ;
     elif [ "`cat $OLD | md5sum`" = "`cat $NEW | md5sum`" ]; then
       # toss the redundant copy
-      rm $NEW
+      rm -f $NEW
     fi
     # Otherwise, we leave the .new copy for the admin to consider...
   done
@@ -36,3 +38,17 @@ for file in etc/profile.d/libgtk+.csh.new etc/profile.d/libgtk+.sh.new ; do
   fi
   config $file
 done
+
+
+# Add gdk modules query update to start to rc.local
+if ! grep "gdk-pixbuf-query-loaders" etc/rc.d/rc.local >/dev/null 2>&1; then
+  cat <<EOF >>etc/rc.d/rc.local
+
+# GNOME applications like this file kept updated:
+if [ -x /usr/bin/gdk-pixbuf-query-loaders ]; then
+  echo "Updating gdk-pixbuf-loaders:  /usr/bin/gdk-pixbuf-query-loaders &"
+   /usr/bin/gdk-pixbuf-query-loaders > /etc/gtk-2.0/gdk-pixbuf.loaders &
+fi;
+
+EOF
+fi
